Move some environment functions out of the C lib
